+++ src/Presentation/wwwroot/js/home/charts/chartTypes/doughnutChart.js
