+++ src/Presentation/wwwroot/js/utils/formatDate.js
@@ -1,0 +1,7 @@
+export function formatDate(dateInput) {
+    const date = new Date(dateInput);
+    const day = date.getDate();
+    const month = date.getMonth() + 1;
+    const year = date.getFullYear();
+    return `${day}/${month}/${year}`;
+}
