+++ src/Presentation/wwwroot/js/utils/setCurrentDateInSelectByDate.js
@@ -15,6 +15,6 @@ if (createdAtFilter) {
     createdAtFilter.value = setCurrentDateInSelectByDate();
 } else {
     // Set current date to initial and final dates filter on Views/Home/Home.cshtml
-    document.getElementById('initialDate').value = setCurrentDateInSelectByDate();
-    document.getElementById('finalDate').value = setCurrentDateInSelectByDate();
+    // document.getElementById('initialDate').value = setCurrentDateInSelectByDate();
+    // document.getElementById('finalDate').value = setCurrentDateInSelectByDate();
 }
